+++ models/contacts.json
@@ -24,12 +24,6 @@
     "phone": "(692) 802-2949"
   },
   {
-    "id": "5",
-    "name": "Cyrus Jackson",
-    "email": "nibh@semsempererat.com",
-    "phone": "(501) 472-5218"
-  },
-  {
     "id": "6",
     "name": "Abbot Franks",
     "email": "scelerisque@magnis.org",
@@ -64,5 +58,11 @@
     "name": "QWeewq QWeewq",
     "email": "newnewnewnew@gmail.com",
     "phone": "915299512959"
+  },
+  {
+    "id": "5",
+    "name": "Chan Gan",
+    "email": "idaididaidia@rst.com",
+    "phone": "5554555555"
   }
 ]
